+++ lib/puree/xml_extractor/organisational_unit.rb
@@ -43,8 +43,14 @@ module Puree
 
       # @return [Puree::Model::OrganisationalUnitHeader, nil]
       def parent
+        multiple_parents = parents
+        multiple_parents.empty? ? nil : multiple_parents.first
+      end
+
+      # @return [Array<Puree::Model::OrganisationalUnitHeader>]
+      def parents
         xpath_result = xpath_query '/parents/parent'
-        Puree::XMLExtractor::Shared.organisation_header xpath_result
+        Puree::XMLExtractor::Shared.organisation_multi_header xpath_result if xpath_result
       end
 
       # @return [Array<String>]